+++ src/components/CalculatorSelectionOverlay/CalculatorSelectionOverlay.vue
@@ -5,15 +5,28 @@
         <!-- The .stop below prevents the click event from bubbling, resulting in the
         overlay only being closed when the modal-mask is clicked directly -->
         <div class="modal-container" v-on:click.stop="overlayClicked">
-          <div class="modal-body" style="display: flex; flex-direction: column; height: 100%;">
+          <div class="modal-body" style="display: flex; flex-direction: row; height: 100%;">
+            <!------------------------>
+            <!-- CATEGORY TREE VIEW -->
+            <!------------------------>
+            <div style="width: 200px; text-align: left; border-right: 1px solid #cdcdcd;">
+              <tree-view :data="treeData" v-on:clicked="categoryClicked" />
+            </div>
+            <!-- The flex: 1 below makes the grid/search take up the remaining horizontal space
+             once the category filter has been assigned space on the left hand-side -->
+          <div style="flex: 1; display: flex; flex-direction: column; height: 100%;">
+            <!---------------------->
             <!-- SEARCH CONTAINER -->
+            <!---------------------->
             <div id="search-container">
               <span style="padding-right: 5px;">Search</span>
               <input v-model="searchText" style="width: 400px; height: 25px;">
             </div>
+            <!---------------------------------->
             <!-- GENERATE CALCULATOR PREVIEWS -->
+            <!---------------------------------->
             <transition-group name="list" tag="div" class="preview-grid">
-              <CalcPreview v-for="item in $store.state.core.filteredAvailableCalcs"
+              <CalcPreview v-for="item in $store.state.core.calcsFilteredByCategoryAndSearch"
                            class="list-complete-item"
                            :title='item.displayName'
                            :description="item.description"
@@ -23,9 +36,10 @@
               </CalcPreview>
             </transition-group>
           </div>
+          </div>
+          </div>
         </div>
       </div>
-    </div>
   </transition>
 </template>
 
@@ -39,19 +53,73 @@
       CalcPreview
     },
     data: function () {
-      return {}
+      return {
+      }
     },
     computed: {
       searchText: {
         get () {
-          return this.$store.state.searchText
+          return this.$store.state.core.searchText
         },
         set (value) {
           this.$store.dispatch('setSearchText', value)
         }
+      },
+      treeData () {
+        var output = {
+          'name': 'root',
+          'selected': false,
+          'children': []
+        }
+
+        console.log(this.$store)
+        var self = this
+        this.$store.state.core.availableCalcs.map(function (calc) {
+          // console.log('calc =')
+          // console.log(calc)
+          self.addCategoriesToTree(calc.category, output)
+          // console.log('After adding category elements for 1 calculator, output =')
+          // console.log(output)
+        })
+        return output
       }
     },
     methods: {
+      addCategoriesToTree (categories, treeNode) {
+        // console.log('addCategoriesToTree() called with categories =')
+        console.log(categories)
+
+        // Copy array, we are going to modify it, and we don't want to touch
+        // the original!
+        categories = categories.slice()
+        // console.log(', treeNode =')
+        // console.log(treeNode)
+
+        var inTree = false
+        treeNode.children.map(function (childNode) {
+          if (childNode.name === categories[0]) {
+            inTree = true
+          }
+        })
+
+        // Add first category element if it doesn't already exist
+        if (!inTree) {
+          console.log('Category ' + categories[0] + ' not found in tree.')
+          var newTreeNode = {
+            'name': categories[0],
+            'selected': false,
+            'children': []
+          }
+          treeNode.children.push(newTreeNode)
+        }
+
+        // Remove first category element
+        categories.splice(0, 1)
+
+        if (categories.length > 0) {
+          this.addCategoriesToTree(categories, treeNode.children[treeNode.children.length - 1])
+        }
+      },
       borderClicked (event) {
         this.$store.commit('showCalculatorSelectionOverlay', {
           trueFalse: false
@@ -60,10 +128,26 @@
       overlayClicked (event) {
         // Do nothing, this handler exists purely to swallow event
         // (event does not bubble because of .stop modifier in HTML)
+      },
+      itemClick (node) {
+        console.log(node.model.text + ' clicked !')
+      },
+      categoryClicked (payload) {
+        console.log('categoryClicked() called. payload =')
+        console.log(payload)
+
+        if (payload.isSelected) {
+          this.$store.dispatch('setSelCategory', payload.category)
+        } else {
+          this.$store.dispatch('setSelCategory', [])
+        }
       }
     },
     watch: {},
     mounted () {
+      if (!this.$store.state.core.calcsFilteredByCategoryAndSearch) {
+        throw Error('this.$store.state.core.calcsFilteredByCategoryAndSearch was null.')
+      }
     }
   }
 </script>